requirements.txt
fastapi==0.70.1
uvicorn==0.18.3
PyYAML>=6.0
opencv_python>=4.5.0
pandas>=1.4.4
requests>=2.28.1
Pillow>=9.2.0
tqdm>=4.64.1
matplotlib>=3.6.0
seaborn>=0.11.0
scipy>=1.4.1
psutil>=5.9.2
pydantic>=1.10.2
http://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp310-cp310-linux_x86_64.whl
http://download.pytorch.org/whl/cpu/torchvision-0.13.1%2Bcpu-cp310-cp310-linux_x86_64.whl